MAJOR: making several breaking changes to the API and types.
