+++ src/api/risc0/groth16.ts
@@ -1,4 +1,4 @@
-import { Groth16ComputationalPlan } from '../../compute/plans/groth16/index.js';
+import { Risc0Groth16ComputationalPlan } from '../../compute/plans/risc0/groth16.js';
 import { Logger } from 'esm-iso-logger';
 import { ApiMethod } from '../methodDecorator.js';
 import {
@@ -21,7 +21,7 @@ const fromRisc0Object = (obj: Risc0ToGroth16Input): Risc0ToGroth16Input => {
   };
 };
 
-export const performRisc0ToGroth16 = ApiMethod<
+export const performRisc0Groth16 = ApiMethod<
   Risc0ToGroth16Input, // TInput (what executor expects)
   typeof risc0ArgsKeys, // TKeys (what arguments mode expects to be provided) performRisc0ToGroth16.fromArgs(risc0_proof, raw_vk)
   Risc0ToGroth16Input // TObject (what object mode expects as a single object) performRisc0ToGroth16.fromObject({} as Risc0ToGroth16Input)
@@ -31,5 +31,5 @@ export const performRisc0ToGroth16 = ApiMethod<
   risc0ObjKeys
 )(async (executor, input) => {
   logger.log('Performing Risc0 to Groth16 conversion...');
-  return executor.execute(new Groth16ComputationalPlan(), input);
+  return executor.execute(new Risc0Groth16ComputationalPlan(), input);
 });
